backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary>=2.9.9
alembic==1.14.0

# Authentication
firebase-admin==6.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud Storage
google-cloud-storage==2.14.0

# HTTP client for Apple IAP verification
httpx==0.28.1

# Environment and validation
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
